move video stuff to separate module
