+++ moseq2_ephys_sync/video/basler_bonsai.py
@@ -9,41 +9,7 @@ import moseq2_ephys_sync.sync as sync
 import moseq2_ephys_sync.util as util
 
 
-def basler_bonsai_workflow(base_path, save_path, num_leds, leds_to_use, led_blink_interval, bonsai_spec='header', timestamp_jump_skip_event_threshhold=1):
-    """
-    Workflow to get codes from bonsai outputted txt file. 
 
-    Bonsai is a "reactive" language, meaning it generates output only when it gets input.
-    In this case, the input that it's yoked to is the Basler camera, recording at ~120 Hz (with 1 ms exposures).
-    So, there should be as many rows in the text file as there are frames in the movie. If this is true, everything is easy --
-    just assign each frame the corresponding timestamp, and sync the timestamps.
-    The timestamps are in microseconds by default.
-    
-    Inputs:
-        base_path (str): path to the .txt file
-        num_leds (int): expects 4
-        leds_to_use (list): eg ['1', '2', '3', '4']. Must contain 4.
-        led_blink_interval: sets an upper bound (in seconds) on how fast the sync code can change. Useful for noisy vids.
-        bonsai_spec (str): Specifies what the column names should be in the data that gets read in. Current option is "default" (time, img num, led1, led2, led3, led4) or "header" (use header in file)
-        timestamp_jump_skip_event_threshhold (int): if there is a jump in timestamps larger than this (in seconds), skip any artifactual "event" that might arise because of it.
-    """
-    print('Doing bonsai basler workflow...')
-    assert num_leds == len(leds_to_use)
-    txt_colnames, txt_dtypes = get_col_info(bonsai_spec)
-    txt_data = load_txt_data(base_path, txt_colnames, txt_dtypes, file_glob='*.csv')
-    bonsai_timestamps = txt_data.time / 1e9  # these are in MICROseconds, convert to seconds
-
-    led_names = ['led1', 'led2', 'led3', 'led4']
-    led_list = []
-    for idx in leds_to_use:
-        led_list.append(txt_data[led_names[int(idx) - 1]])
-
-    bonsai_events = list_to_events(bonsai_timestamps, led_list, tskip=timestamp_jump_skip_event_threshhold)
-    bonsai_codes, _ = sync.events_to_codes(bonsai_events, nchannels=num_leds, minCodeTime=led_blink_interval-1) 
-    bonsai_codes = np.asarray(bonsai_codes)
-    bonsai_codes[:,0] = bonsai_codes[:,0]
-
-    return bonsai_codes, bonsai_timestamps
